Checkpoint #2. Did a lot. Added functions to edit and add items.Changed some UI styles. More automation

+++ intern_app/custom_admin.py
@@ -1,21 +1,32 @@
-from django.shortcuts import render, redirect, get_object_or_404
+from django.shortcuts import render, redirect, get_object_or_404 ,HttpResponseRedirect
 from django.contrib import admin
+from django.contrib.auth import authenticate, login, logout, get_user_model, update_session_auth_hash
+from django.contrib.auth.models import User
 from django.apps import apps
 from django.urls import reverse
 from django.forms import modelform_factory
 from django.http import Http404
-from django import forms
+from django.contrib import messages 
+from .models import Categories
+from django.contrib.auth.decorators import login_required
+from django.core.exceptions import FieldDoesNotExist
+
 # Burada tüm custom-admin sayfasının fonksiyonları yer alacaktır.,
-def model_silme(request,model_name,model_id ):
-    tum_modeller = apps.get_model("intern_app",model_name)
-    model_objects = tum_modeller.objects.get(id = model_id)
+User = get_user_model
+@login_required
+def model_silme(request, model_name,model_id):
+    if request.user.is_superuser:
+        tum_modeller = apps.get_model("intern_app",model_name)
+        model_objects = tum_modeller.objects.get(id = model_id)
 
-    model_objects.delete()
 
-    return redirect(request.META.get('HTTP_REFERER'))
+        model_objects.delete()
+
+        return redirect(request.META.get('HTTP_REFERER'))
+@login_required
 # Obje güncelleme fonksiyonu
-
 def dynamic_model_item_update(request,app_label,model_name,object_id):
+   if request.user.is_superuser:
     
     try:
         model= apps.get_model(app_label,model_name)
@@ -25,7 +36,8 @@ def dynamic_model_item_update(request,app_label,model_name,object_id):
     # Model ve nesnesini alıyoruz.
     model_object = get_object_or_404(model,id=object_id)
     verbose_name = model._meta.verbose_name_plural
-
+    obj_names = [model_object]  # obj_names artık sadece tek bir öğe içeriyor
+    get_names = [{"obj": model_object}]  # Aynı öğe sadece bir kez ekleniyor
     #Formu oluşturuyoruz.
     modelForm = modelform_factory(model,fields="__all__")
 
@@ -35,20 +47,25 @@ def dynamic_model_item_update(request,app_label,model_name,object_id):
         form = modelForm(request.POST,request.FILES,instance=model_object)
         if form.is_valid(): # Eğer form geçerliyse formu kaydet
             form.save()
+            messages.success(request, 'Başarıyla güncellendi!')
             return redirect(reverse("intern_app:update_model",kwargs={"app_label":app_label,"model_name":model_name,"object_id":model_object.id}))
     else:
             form = modelForm(instance=model_object) 
         
     context = {
-     "model_object":model_object,
-     "model_name":model_name,
-     "form":form,
-     "verbose_name":verbose_name,
+        "model_object":model_object,
+        "model_name":model_name,
+        "get_names":get_names,
+        "obj_names":obj_names,
+        "form":form,
+        "verbose_name":verbose_name,
     }
     return render(request, "intern_app/custom_admin_update_item.html", context)
-
+    
+@login_required
 # Obje ekleme fonksiyonu
 def dynamic_model_item_add(request,app_label, model_name):
+   if request.user.is_superuser:
   
     
     # Modeli getir
@@ -66,22 +83,51 @@ def dynamic_model_item_add(request,app_label, model_name):
 
     if request.method == "POST":
         form = ModelForm(request.POST, request.FILES)
+        is_sub_cat = request.POST.get("is_subs") == "on" # Onay kutusu seçilmiş mi ?
+        parent_category_id = request.POST.get("parent_category") # Seçilen üst kategori
+
         if form.is_valid():
-            form.save()
+            new_obj = form.save(commit=False)
+            new_obj.is_sub_cat = is_sub_cat
+            new_obj.save()
+            if is_sub_cat:
+                if not parent_category_id:
+                    form.add_error(None,"Alt Kategori olarak belirlediğiniz öğenin bir üst kategorisi olmalıdır!")
+                else:
+                    try:
+                        parent_cat = Categories.objects.get(id=parent_category_id)
+                        parent_cat.sub_categories.add(new_obj)
+                        new_obj.save()
+                        messages.success(request, 'Başarıyla eklendi!')
+                        #new_obj.sub_categories.add(parent_cat)
+                        #new_obj.parent_categories = Categories.objects.get(id=parent_category_id)
+                    except Categories.DoesNotExist:
+                        form.add_error =(None,"Seçilen üst kategori bulunamadı.")
+            else:
+                new_obj.save()
+         
+            
+            
+            
             return redirect(reverse("intern_app:add_model", kwargs={"app_label": app_label, "model_name": model_name}))  
     else:
         form = ModelForm()
+
+    categories = Categories.get_main_categories  # Sadece ana kategoriler
 
     context = {
         "model_objects": model_objects,
         "verbose_name": verbose_name,
         "form": form,
+        "categories": categories,
         "model_name": model_name
     }
 
     return render(request, "intern_app/custom_admin_add_item.html", context)
-
+@login_required
 def admin_links(request):
+   if request.user.is_superuser:
+
     # Modelleri koymak için liste
     models = []
     # Uygulamamızdaki modelleri model isminde alıyoruz.
@@ -93,35 +139,46 @@ def admin_links(request):
             verbose_name = model._meta.verbose_name_plural # Modelin çoğul adı
             # Aldığımız verileri listeye ekliyoruz.
             models.append({"app_label": app_label, "model_name": model_name,"verbose_name":verbose_name})
-
-    return render(request,"intern_app/custom-admin.html", {"models": models})
-
-def admin_links_pages(request, app_label, model_name):
-    model_class = apps.get_model(app_label, model_name)
-    model_objects = model_class.objects.all()
-    verbose_name = model_class._meta.verbose_name_plural
-    model_name = model_class._meta.model_name
-    app_label = model_class._meta.app_label
-    
-    model_data = []
-    for obj in model_objects:
-        category_type = "Alt Kategori" if hasattr(obj, 'parent_categories') and obj.parent_categories.exists() else "Ana Kategori"
-
-        model_data.append({
-            "obj": obj,
-            "category_type": category_type,
-            
-        })
-
+    verbose_name = model._meta.verbose_name_plural
+    obj_names = [models]  # obj_names artık sadece tek bir öğe içeriyor
+    get_names = [{"obj": models}]  # Aynı öğe sadece bir kez ekleniyor
     context = {
-        "model_objects": model_objects,
-        "model_class": model_class,
-        "model_name":model_name,
-        "model_data":model_data,
-        "app_label":app_label,
-        "verbose_name": verbose_name,
-        "category_header":"Kategori Türü",
-        
+        "models":models,
+        "obj_names":obj_names,
+        "get_names":get_names
     }
+    return render(request,"intern_app/custom-admin.html", context)
+   raise Http404("Yetkiniz yok.")
+@login_required
+def admin_links_pages(request, app_label, model_name):
+    if request.user.is_superuser:
 
-    return render(request, 'intern_app/custom_admin_pages.html', context)
+        model_class = apps.get_model(app_label, model_name)
+        model_objects = model_class.objects.all()
+        verbose_name = model_class._meta.verbose_name_plural
+        model_name = model_class._meta.model_name
+        app_label = model_class._meta.app_label
+        
+        model_data = []
+        for obj in model_objects:
+            category_type = "Alt Kategori" if hasattr(obj, 'parent_categories') and obj.parent_categories.exists() else "Ana Kategori"
+
+            model_data.append({
+                "obj": obj,
+                "category_type": category_type,
+                
+            })
+
+
+        context = {
+            "model_objects": model_objects,
+            "model_class": model_class,
+            "model_name":model_name,
+            "model_data":model_data,
+            "app_label":app_label,
+            "verbose_name": verbose_name,
+            "category_header":"Kategori Türü",
+            
+        }
+
+        return render(request, 'intern_app/custom_admin_pages.html', context)
